[DBC] Final Fix for TickTime

+++ Parse_SpellTickTime.py
@@ -19,14 +19,14 @@ with open(os.path.join(generatedDir, 'SpellEffect.csv')) as csvfile:
     reader = csv.DictReader(csvfile, escapechar='\\')
     ValidRows = []
     for row in reader:
-        if not float(row['amplitude']) == 0:
+        if not int(row['amplitude']) == 0:
             ValidRows.append(row)
     with open(os.path.join(parsedDir, 'TickTime.lua'), 'w', encoding='utf-8') as file:
         file.write('AethysCore.Enum.TickTime = {\n')
         iMax = len(ValidRows)-1
         for i, row in enumerate(ValidRows):
             if i == iMax:
-                file.write('  [' + row['id_parent'] + '] = {' + str(int(float(row['amplitude']))) + ', ' + ('false' if row['id_mechanic'] == "15" else 'true') + '}\n')
+                file.write('  [' + row['id_parent'] + '] = {' + str(int(row['amplitude'])) + ', ' + ('false' if row['id_mechanic'] == "15" else 'true') + '}\n')
             else:
-                file.write('  [' + row['id_parent'] + '] = {' + str(int(float(row['amplitude']))) + ', ' + ('false' if row['id_mechanic'] == "15" else 'true') + '},\n')
+                file.write('  [' + row['id_parent'] + '] = {' + str(int(row['amplitude'])) + ', ' + ('false' if row['id_mechanic'] == "15" else 'true') + '},\n')
         file.write('}\n')
